proce number formate update

+++ app/Http/Resources/Api/Subscription/SubscriptionPlanResource.php
@@ -15,7 +15,7 @@ class SubscriptionPlanResource extends JsonResource {
         return [
             'id'                => $this->id,
             'subscription_plan' => $this->subscription_plan,
-            'price'             => (float) $this->price,
+            'price'             => (double) $this->price,
             'billing_cycle'     => $this->billing_cycle,
             'touch_points'      => $this->touch_points === null ? 'Unlimited Touchpoints' : $this->touch_points,
             'has_ads'           => (bool) $this->has_ads,
